transform from barycentric to 3d coordinates inside the kernel function
Note: this is producing odd results. Im led to believe that the UV coords are incorrect
even though saving them directly in an image seemed fine before.

I will leave this branch where it is now, and move back to the idea of
having a special kernel shading function for this. Though I'm probably
going to run into the same issue there ...