Cycles-Bake: Tangent Normal space mode
TANGENT gives no crash and bluish results, so it seems correct.
There are some differences from BI and Cycles, I'll have to test further
to see which one is correct.

Apart from that there is a chance the swizzle is wrong.
In other words, the directions may be swapped, or inverted.
It will be an easy fix though. I just need to put time on testing this
against other applications and simple test files.

* Refactored calculateTriTessFaces to store tangent.
* Re-set neutral color for WORLD/OBJECT maps to be 0,0,0 from Cycles