Cycles-Bake: Support for multiple high-poly objects selected at once
When using "selected to active" the user can now have more than one
highpoly object for a single lowpoly object.

Code note: instead of implementing a new BHV method to create a tree
with multiple meshes, we are casting the ray to the individual trees
(for each highpoly object) and comparing the distances. It could be
optimized with the BVH change, but it will do for now.

Also note: the same method can be used to implement 'Blockers'.