Bake-API: new approach for cage
For users:
The cage now controls the direction and the origin of the
rays casted to the highpoly objects. The direction is a ray from the
point in the cage mesh to the equivalent point to the base mesh. That
means the face normals are entirely ignored when using a cage object.

The option cage_extrusion only works on top of the base mesh, and it's
ignored when using cage objects. When not using cage objects the normals
of the faces are taken into consideration, which means the user should
use Split Modifiers in some cases (or cages) to control the normal
interpolation (or add more loops in the extremety of the faces). See
T40023

For developers:
When using an object cage the ray is calculated from the cage mesh to
the base mesh. It uses the barycentric coordinate from the base mesh UV,
so we expect both meshes to have the same primitive ids (which won't be
the case if the cage gets edited in a destructive way).

Also, now cage_extrusion only affects non-cage baking. So the user has
the option to use a cage mesh, or to extrude the base mesh.

Also, we now accept the Edge Split modifier in the base mesh when using
cage_extrusion (related to T40023).

More on cages:
http://wiki.polycount.com/NormalMap/#Working_with_Cages

Reviewers: campbellbarton, brecht

CC: adriano, metalliandy

Differential Revision: https://developer.blender.org/D547